requirements_rag.txt
requests>=2.31.0
# google-generativeai>=0.3.0 (Removed to fix pydantic-core binary error)
# chromadb>=0.4.0 (Removed to fix binary loading errors)
numpy>=1.24.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pymupdf>=1.24.0
